changed PI Library to take Float for DeltaTime Changed timing to be more accurate in main loop. Still need to switch to micros(). Left that for Tridge. This gets rid of six floating point calcs *250. Nice!

+++ libraries/APM_PI/APM_PI.h
@@ -77,7 +77,7 @@ public:
 	///
 	/// @returns		The updated control output.
 	///
-	long 	get_pi(int32_t error, uint16_t dt);
+	long 	get_pi(int32_t error, float dt);
 
 	/// Reset the PI integrator
 	///
